removed menu code to create JSON
This is to reduce the amount of html code and create the menu using a JSON array instead. The JSON data will be used to dynamically populate the menu in HTML using Javascript & jQuery. This will make it much easier to update if new menu dishes were to be added in the future.