separated menu array into nested arrays
This is so that I can better organize the layout of the menu by separating the main dishes into two columns instead of one, and also putting the desserts and drink items into their own section in the html. I used different for loops to interate through the nested arrays.